+++ QRCodeReader/Classes/Views/QRCode/QRCodeReaderView.swift
@@ -1,0 +1,61 @@
+//
+//  Copyright (c) 2019 Touch Instinct
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the Software), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED AS IS, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import UIKit
+
+open class QRCodeReaderView: BaseReaderView {
+    
+    private weak var reader: QRCodeReader?
+
+    // MARK: - Public Methods
+
+    public func setReader(_ reader: QRCodeReader) {
+        self.reader = reader
+        self.reader?.onUpdateRectOfInterest = updateRectOfInterestBasedOnFocusView
+        
+        cameraView.layer.sublayers?.forEach {
+            $0.removeFromSuperlayer()
+        }
+
+        cameraView.layer.insertSublayer(reader.previewLayer, at: 0)
+    }
+
+    public func updateRectOfInterestBasedOnFocusView() {
+        guard let reader = reader, let focusView = overlay.focusView else {
+            return
+        }
+
+        let focusViewRect = convert(focusView.frame, from: focusView)
+        let interestRect = reader.previewLayer.metadataOutputRectConverted(fromLayerRect: focusViewRect)
+
+        reader.metadataOutput.rectOfInterest = interestRect
+    }
+    
+    override open func layoutSubviews() {
+        super.layoutSubviews()
+        
+        reader?.previewLayer.frame = bounds
+        overlay.frame = bounds
+        
+        updateRectOfInterestBasedOnFocusView()
+    }
+}
